Move Base64Encoder to src/anyp/
