+++ src/Makefile.am
@@ -22,7 +22,6 @@ SUBDIRS = \
 	time \
 	debug \
 	base \
-	base64 \
 	anyp \
 	helper \
 	dns \
@@ -511,7 +510,6 @@ squid_LDADD = \
 	http/libhttp.la \
 	dns/libdns.la \
 	base/libbase.la \
-	base64/libbase64.la \
 	libsquid.la \
 	fs/libfs.la \
 	DiskIO/libdiskio.la \
@@ -993,7 +991,8 @@ nodist_tests_testBase64Encoder_SOURCES = \
 tests_testBase64Encoder_LDADD = \
 	sbuf/libsbuf.la \
 	base/libbase.la \
-	base64/libbase64.la \
+	anyp/libanyp.la \
+	$(top_builddir)/lib/libmiscencoding.la \
 	$(LIBCPPUNIT_LIBS) \
 	$(COMPAT_LIB) \
 	$(XTRA_LIBS) \
@@ -2279,7 +2278,6 @@ tests_testCacheManager_LDADD = \
 	eui/libeui.la \
 	icmp/libicmp.la \
 	log/liblog.la \
-	base64/libbase64.la \
 	format/libformat.la \
 	$(REPL_OBJS) \
 	$(ADAPTATION_LIBS) \
